requirements.txt
panda3d
pyzmq
